(possibly) enlarge dequeue() to give us better statistics
It happened on Fedora 23 x86_64 that if the next() was called with
constant window of n=8196 units (in Bytes, default with
MultipartEncoderMonitor) -- some cache/buffer on the road caused
that the sum of deltas started to be highly inaccurate (bar shown
like 40MB instead of 1MB _most_ of the time).  This could be
solved by higher sma_window;  but we would have to rather large
numbers there and that would result in high complexity.

So store into deque() also "actual" time() and count all units in
window in separate variable (self._in_window) to avoid sum()
complexity issues.  That way we may count:

    (actual_time - last_item_time) / items_in_window

This would allow us to raise the value of sma_window.